🎨 style(member): reformat code, code sorting
