+++ src/main/java/com/ptjcoding/nbcampspringnewsfeed/domain/member/service/MemberService.java
@@ -3,14 +3,15 @@ package com.ptjcoding.nbcampspringnewsfeed.domain.member.service;
 import com.ptjcoding.nbcampspringnewsfeed.domain.common.dto.CommonResponseDto;
 import com.ptjcoding.nbcampspringnewsfeed.domain.member.dto.LoginRequestDto;
 import com.ptjcoding.nbcampspringnewsfeed.domain.member.dto.SignupRequestDto;
+import com.ptjcoding.nbcampspringnewsfeed.domain.member.service.dto.MemberResponseDto;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.http.ResponseEntity;
-import org.springframework.transaction.annotation.Transactional;
 
 public interface MemberService {
 
-  ResponseEntity<CommonResponseDto<Void>> signup(
+  ResponseEntity<CommonResponseDto<MemberResponseDto>> signup(
       SignupRequestDto signupRequestDto);
 
-  ResponseEntity<CommonResponseDto<Void>> login(LoginRequestDto loginRequestDto, HttpServletResponse response);
+  ResponseEntity<CommonResponseDto<Void>> login(LoginRequestDto loginRequestDto,
+      HttpServletResponse response);
 }
